Moved Lang for commands lang -> lang.command
