+++ src/main/java/com/hm/achievement/command/HelpCommand.java
@@ -1,9 +1,9 @@
 package com.hm.achievement.command;
 
 import com.hm.achievement.AdvancedAchievements;
-import com.hm.achievement.lang.CmdLang;
-import com.hm.achievement.lang.HelpLang;
 import com.hm.achievement.lang.Lang;
+import com.hm.achievement.lang.command.CmdLang;
+import com.hm.achievement.lang.command.HelpLang;
 import com.hm.mcshared.particle.PacketSender;
 import org.bukkit.ChatColor;
 import org.bukkit.command.CommandSender;
