+++ src/main/java/com/hm/achievement/command/InfoCommand.java
@@ -1,8 +1,8 @@
 package com.hm.achievement.command;
 
 import com.hm.achievement.AdvancedAchievements;
-import com.hm.achievement.lang.InfoLang;
 import com.hm.achievement.lang.Lang;
+import com.hm.achievement.lang.command.InfoLang;
 import org.apache.commons.lang3.StringEscapeUtils;
 import org.bukkit.Bukkit;
 import org.bukkit.ChatColor;
