+++ src/main/java/com/hm/achievement/command/StatsCommand.java
@@ -3,8 +3,8 @@ package com.hm.achievement.command;
 import com.hm.achievement.AdvancedAchievements;
 import com.hm.achievement.category.MultipleAchievements;
 import com.hm.achievement.category.NormalAchievements;
-import com.hm.achievement.lang.CmdLang;
 import com.hm.achievement.lang.Lang;
+import com.hm.achievement.lang.command.CmdLang;
 import com.hm.mcshared.particle.ParticleEffect;
 import org.bukkit.ChatColor;
 import org.bukkit.command.CommandSender;
