+++ src/main/java/com/hm/achievement/lang/command/InfoLang.java
@@ -1,4 +1,6 @@
-package com.hm.achievement.lang;
+package com.hm.achievement.lang.command;
+
+import com.hm.achievement.lang.Lang;
 
 public enum InfoLang implements Lang {
 	DESCRIPTION("Description:"),
