+++ src/test/java/com/hm/achievement/lang/HelpLangTest.java
@@ -1,13 +1,13 @@
 package com.hm.achievement.lang;
 
+import com.hm.achievement.lang.command.HelpLang;
 import org.junit.Test;
 
 import java.util.Arrays;
-import java.util.HashSet;
 import java.util.Set;
 import java.util.stream.Collectors;
 
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertEquals;
 
 public class HelpLangTest {
